fixed imports, issues from code refactoring
