+++ src/components/SideBar/index.jsx
@@ -1,6 +1,7 @@
 import "./index.css";
 
 import { Link, useNavigate } from "react-router-dom";
+import React, { useState } from "react";
 
 import AccountCircleIcon from "@mui/icons-material/AccountCircle";
 import AddCircleOutlineIcon from "@mui/icons-material/AddCircleOutline";
@@ -10,9 +11,7 @@ import FavoriteBorderIcon from "@mui/icons-material/FavoriteBorder";
 import HomeIcon from "@mui/icons-material/Home";
 import ImgUpload from "../ImgUpload";
 import NotificationsIcon from "@mui/icons-material/Notifications";
-import React from "react";
 import { auth } from "../../lib/firebase";
-import { useState } from "react";
 
 function SideBar() {
   const navigate = useNavigate();
